Add files via upload
My name is Daniele, I'm looking for funds to help treat my mother with Alzheimer's in the best possible way, I'm also evaluating the use of the drug Aducanumab produced by Biogen, but even there the cost is prohibitive and illegal selling in Europe.
In general, I want to organize the best 24/7 care for my mother, I try to raise some money so that her illness is as calm as possible.
Of these Tokens I only keep 3.33% for me and the guy who is helping me create it, all the rest is freely available and a small part for the first few years will be used to minimize speculation or any attempt to Pump and Down. (forbidden and immoral) and pay taxes because here in Italy they have to be declared to the state and a part will go to the tax authorities.
If I could raise a significant amount of capital above what I needed, I would use it to help other struggling families like ours.
Just this.